requirements.txt
# Core dependencies
openai==1.12.0
python-dotenv==1.0.0
SpeechRecognition==3.10.1
pyttsx3==2.90
edge-tts==6.1.9
openai-whisper==20231117
numpy==1.24.3
pyaudio==0.2.14
google-generativeai==0.3.1
requests==2.31.0

# Computer Vision
opencv-python==4.9.0.80
tensorflow==2.15.0
face-recognition==1.3.0
dlib==19.24.2

# Web and API
beautifulsoup4==4.12.2
selenium==4.16.0
webdriver-manager==4.0.1
instaloader==4.10.1
pywhatkit==5.4
python-telegram-bot==20.7

# Database
sqlalchemy==2.0.25
pymongo==4.6.1

# GUI
PyQt6==6.6.1
customtkinter==5.2.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3
pillow==10.1.0
pandas==2.1.4
pyautogui==0.9.54
psutil>=5.9.0
pywin32==306 